+++ src/app/components/shared/stepper/stepper.component.scss
@@ -21,8 +21,7 @@
   .lines {
     align-items: center;
     display: flex;
-    // TODO This does not line up the line with the center of the steps universally
-    height: 60%;
+    height: 50%;
     position: absolute;
     top: 0%;
     width: 100%;
